Leveling system implemented

+++ app/src/main/java/com/example/ubermensch/models/Experience.kt
@@ -2,5 +2,5 @@ package com.example.ubermensch.models
 
 class Experience(
     var Level: Int? = 1,
-    var XP: Long?=0
+    var XP: Double? =0.0
 )
